Minor adjustments
Nothing special, just moved some stuff around for better organization and I also added the remaining guns to the stations. They need some more adjustment as well, but the general functionality is working. I need to update the UI menu as well when the camera gets switched. I will probably fix that  in the morning.